+++ packages/template-retail-react-app/app/utils/sfdc-user-agent-utils.js
