+++ packages/template-retail-react-app/app/utils/sfdc-user-agent-utils.test.js
@@ -5,9 +5,9 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import {generateSfdcUserAgent} from '@salesforce/retail-react-app/../../app/utils/user-agent-utils'
+import {generateSfdcUserAgent} from './sfdc-user-agent-utils'
 
-describe('user-agent-utils', () => {
+describe('sfdc-user-agent-utils', () => {
     describe('generateSfdcUserAgent', () => {
         test('should generate correct sfdc_user_agent header value', () => {
             const userAgent = generateSfdcUserAgent()
@@ -44,6 +44,40 @@ describe('user-agent-utils', () => {
             const userAgent = generateSfdcUserAgent()
             if (userAgent.includes('unknown')) {
                 expect(userAgent).toBe('pwa-kit-react-sdk@unknown commerce-sdk-react@unknown')
+            }
+        })
+
+        test('should warn and return fallback when package.json cannot be read', () => {
+            const consoleWarnSpy = jest.spyOn(console, 'warn').mockImplementation(() => {})
+            
+            // Mock the module system to simulate error case
+            const originalModule = require.cache[require.resolve('../../package.json')]
+            delete require.cache[require.resolve('../../package.json')]
+            
+            // Force require to fail by temporarily removing the file from cache
+            jest.doMock('../../package.json', () => {
+                throw new Error('Cannot find module')
+            })
+
+            // Re-require the module to get the version with mocked dependency
+            jest.resetModules()
+            const { generateSfdcUserAgent: mockGenerateSfdcUserAgent } = require('./sfdc-user-agent-utils')
+            
+            try {
+                const userAgent = mockGenerateSfdcUserAgent()
+                
+                expect(userAgent).toBe('pwa-kit-react-sdk@unknown commerce-sdk-react@unknown')
+                expect(consoleWarnSpy).toHaveBeenCalledWith(
+                    'Unable to generate sfdc_user_agent header:',
+                    expect.any(Error)
+                )
+            } finally {
+                jest.dontMock('../../package.json')
+                jest.resetModules()
+                if (originalModule) {
+                    require.cache[require.resolve('../../package.json')] = originalModule
+                }
+                consoleWarnSpy.mockRestore()
             }
         })
